broke out datetime component
